+++ src/main/java/insane96mcp/survivalreimagined/module/items/feature/AncientLapis.java
@@ -4,7 +4,6 @@ import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.LoadFeature;
-import insane96mcp.survivalreimagined.data.lootmodifier.ReplaceDropModifier;
 import insane96mcp.survivalreimagined.module.Modules;
 import insane96mcp.survivalreimagined.setup.SRItems;
 import net.minecraft.util.RandomSource;
@@ -15,10 +14,6 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.item.enchantment.EnchantmentInstance;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.storage.loot.predicates.LootItemBlockStatePropertyCondition;
-import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
-import net.minecraftforge.common.data.GlobalLootModifierProvider;
 import net.minecraftforge.event.AnvilUpdateEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.registries.RegistryObject;
@@ -92,17 +87,5 @@ public class AncientLapis extends Feature {
 			EnchantedBookItem.addEnchantment(stack, new EnchantmentInstance(enchantment, level));
 		else
 			stack.enchant(enchantment, level);
-	}
-
-	private static final String path = "ancient_lapis/";
-
-	public static void addGlobalLoot(GlobalLootModifierProvider provider) {
-		provider.add(path + "ancient_lapis_from_deepslate_ore", new ReplaceDropModifier.Builder(
-				new LootItemCondition[]{new LootItemBlockStatePropertyCondition.Builder(Blocks.DEEPSLATE_LAPIS_ORE).build()},
-				Items.LAPIS_LAZULI,
-				ANCIENT_LAPIS.get())
-				.setAmountToReplace(1)
-				.setChances(List.of(0.12f, 0.15f, 0.19f, 0.24f, 0.30f))
-				.build());
 	}
 }