+++ changelog.md
@@ -23,7 +23,10 @@
   * Instead of removing knockback, now attacking without a weapon or spamming has reduced knockback
 * Added Nether Infused Powered Rails
   * 2.5x faster than normal rails
-* Removed Mending and added Cleansing Moss
+* Removed Mending and added Cleansed Lapis
+  * New item that can reset the repair cost of items
+* Added Ancient Lapis
+  * New item crafted from Cleansed Lapis that can increase the level of enchantments over the maximum
 * Slight rework of structure loot (more XP Bottles)
 * Increased Golden Tools base efficiency
 * Sprint jumping is now affected by movement speed